[vm] Fix gcc build
TEST=ci

Cq-Include-Trybots: luci.dart.try:dart-sdk-linux-riscv64-try
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/299940
Commit-Queue: Alexander Markov <alexmarkov@google.com>

+++ runtime/vm/object.cc
@@ -22598,7 +22598,8 @@ bool TypeParameter::IsEquivalent(
     TypeEquality kind,
     FunctionTypeMapping* function_type_equivalence) const {
   TRACE_TYPE_CHECKS_VERBOSE("  TypeParameter::IsEquivalent(%s, %s, kind %d)\n",
-                            ToCString(), other.ToCString(), kind);
+                            ToCString(), other.ToCString(),
+                            static_cast<int>(kind));
   if (ptr() == other.ptr()) {
     TRACE_TYPE_CHECKS_VERBOSE("   - result: true (same types)\n");
     return true;
